+++ NEWS.md
@@ -1,3 +1,5 @@
+# hubAdmin (development version)
+
 # hubAdmin 1.7.0
 
 * update `create_target_metadata_item()`function to add optional additional properties, to match schema version 5.1.0. (#108)